+++ CHANGELOG.md
@@ -5,4 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- The API server can now be run either via `viewerd` or `python -m activity_viewer`.
+- Electron app now lives in the app/ folder.
+- Zero or more file names can be passed as arguments to `viewer visualize`.
+- Validate command is now invoked: `viewer validate FILENAME`, rather than `viewer validate -f FILENAME`. `FILENAME` is now a *required* argument.
 - `blacklist`/`whitelist` fields in settings files are now `exclude`/`include`, respectively.